+++ api/swipr/src/main/java/com/swipr/matcher/Query.java
@@ -3,7 +3,7 @@ package com.swipr.matcher;
 public class Query {
     // User ID of the user that created this query. NOT boxed because we
     // want to at least pretend we care about performance.
-    public final long userId;
+    public final int userId;
 
     // Start and end of the time range (in seconds since epoch) that
     // we are querying for.
@@ -18,7 +18,7 @@ public class Query {
     public static final long BPLATE = 1, COVEL = 2, DE_NEVE = 4, FEAST = 8;
 
     public Query(
-        long userId,
+        int userId,
         long timeRangeStart,
         long timeRangeEnd,
         long priceCents,
